+++ tons.h
@@ -1,6 +1,6 @@
 #ifndef TONS_H
 #define TONS_H
 
-int tons(double primeiro, double segundo);
+int tons(int primeiro, int segundo);
 
 #endif
